TS-238 Refactor: Domain구조 리팩토링 및 LocalDateTime -> LocalTime변경
